SPEED-UP-SEARCH: Modified code to speed up patient search and identifier
The patient searching is implemented in a way isolates names from identifier searches. Moreover
for AMPATH purposes attribute searches are not really important hence skipped. the search string
is assumed to be an identifier if it contain digits otherwise it is name. Moreover if nothing is
is found for any name search is found then an identifier search is performed instead to cover the
case where identifier consist of alphabets only. However this is a stretch for AMPATH because all
identifiers have numbers